+++ app/src/androidTest/java/woowacourse/movie/view/reservation/seat/SeatSelectActivityTest.kt
@@ -13,7 +13,7 @@ import androidx.test.espresso.matcher.ViewMatchers.withText
 import org.junit.Before
 import org.junit.Test
 import woowacourse.movie.R
-import woowacourse.movie.model.MovieTicket
+import woowacourse.movie.model.reservation.MovieTicket
 import woowacourse.movie.view.Extras
 import java.time.LocalDate
