--- app/src/main/java/woowacourse/movie/model/ReservationUIModel.kt
@@ -1,9 +1,0 @@
-package woowacourse.movie.model
-
-data class ReservationUIModel(
-    val movie: Movie,
-    val movieDate: MovieDate,
-    val movieTime: MovieTime,
-    val ticketCount: Int,
-    val theaterName: String,
-)
