+++ app/src/main/java/woowacourse/movie/model/database/DummyMovieDao.kt
@@ -1,7 +1,10 @@
-package woowacourse.movie.model
+package woowacourse.movie.model.database
 
+import woowacourse.movie.model.movie.Movie
+import woowacourse.movie.model.theater.Theater
 import java.time.LocalDate
 import java.time.LocalDateTime
+import kotlin.collections.forEach
 
 object DummyMovieDao : MovieDao {
     override fun getTheaterNames(): List<String> = DummyMovieDB.screenings.keys.toList()
