+++ app/src/main/java/woowacourse/movie/model/database/MovieDatabaseSchema.kt
@@ -1,4 +1,6 @@
-package woowacourse.movie.model
+package woowacourse.movie.model.database
+
+import woowacourse.movie.model.movie.Movie
 
 interface MovieDatabaseSchema {
     val screenings: Map<String, Map<String, List<Int>>>
