+++ app/src/main/java/woowacourse/movie/model/movie/Movie.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.model.movie
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
