+++ app/src/main/java/woowacourse/movie/view/movie/MovieClickListener.kt
@@ -1,6 +1,6 @@
 package woowacourse.movie.view.movie
 
-import woowacourse.movie.model.Movie
+import woowacourse.movie.model.movie.Movie
 
 fun interface MovieClickListener {
     fun onReservationClick(movie: Movie)
