+++ app/src/main/java/woowacourse/movie/view/reservation/seat/SeatSelectContract.kt
@@ -1,8 +1,8 @@
 package woowacourse.movie.view.reservation.seat
 
 import androidx.annotation.StringRes
-import woowacourse.movie.model.MovieTicket
-import woowacourse.movie.model.ReservationInfo
+import woowacourse.movie.model.reservation.MovieTicket
+import woowacourse.movie.model.reservation.ReservationInfo
 import woowacourse.movie.model.seat.index.Col
 import woowacourse.movie.model.seat.index.Row
