+++ app/src/main/java/woowacourse/movie/view/theater/TheaterClickListener.kt
@@ -1,6 +1,6 @@
 package woowacourse.movie.view.theater
 
-import woowacourse.movie.model.TheaterUIModel
+import woowacourse.movie.model.theater.TheaterUIModel
 
 interface TheaterClickListener {
     fun onTheaterClick(theaterUIModel: TheaterUIModel)
