+++ app/src/main/java/woowacourse/movie/view/theater/TheaterDiffUtil.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.view.theater
 
 import androidx.recyclerview.widget.DiffUtil
-import woowacourse.movie.model.TheaterUIModel
+import woowacourse.movie.model.theater.TheaterUIModel
 
 object TheaterDiffUtil : DiffUtil.ItemCallback<TheaterUIModel>() {
     override fun areItemsTheSame(
