+++ app/src/main/res/layout/activity_reservation.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="reservationState"
-            type="woowacourse.movie.model.ReservationUIModel" />
+            type="woowacourse.movie.model.reservation.ReservationUIModel" />
 
         <variable
             name="presenter"
