+++ app/src/main/res/layout/item_movie.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="movie"
-            type="woowacourse.movie.model.Movie" />
+            type="woowacourse.movie.model.movie.Movie" />
 
         <variable
             name="moviesFragment"
