+++ app/src/test/java/woowacourse/movie/domain/model/Fixture.kt
@@ -1,12 +1,12 @@
 package woowacourse.movie.domain.model
 
 import woowacourse.movie.R
-import woowacourse.movie.model.Movie
-import woowacourse.movie.model.MovieDatabaseSchema
-import woowacourse.movie.model.MovieTicket
-import woowacourse.movie.model.ReservationInfo
 import woowacourse.movie.model.Seats
-import woowacourse.movie.model.TheaterUIModel
+import woowacourse.movie.model.database.MovieDatabaseSchema
+import woowacourse.movie.model.movie.Movie
+import woowacourse.movie.model.reservation.MovieTicket
+import woowacourse.movie.model.reservation.ReservationInfo
+import woowacourse.movie.model.theater.TheaterUIModel
 import java.time.LocalDate
 
 val dummyMovie =
