+++ app/src/test/java/woowacourse/movie/domain/model/MovieDateTest.kt
@@ -2,7 +2,7 @@ package woowacourse.movie.domain.model
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import woowacourse.movie.model.MovieDate
+import woowacourse.movie.model.movie.MovieDate
 import java.time.LocalDate
 
 class MovieDateTest {
